// src/app/models/Proyecto.ts
export interface Proyecto {
    id?: number;
    name: string;
    motivo: string;
    date_start: string;
    date_end: string;
    description: string;
    urlImage:string;
    portfolio: string;
    urlImagePort: string;
    presentacion: string;
    urlImageVideo: string;
    redes: string;
    urlImageRedes:string;
    

  }
  